+++ data/src/utils/setting_keywords.rs
@@ -231,6 +231,13 @@ keyword_table! {
         "general.strip_separator" => &["delimiter", "divider"],
         "general.artwork_column_mode" => &["cover art", "album art panel", "sidebar art"],
         "general.artwork_column_stretch_fit" => &["crop", "fill", "aspect ratio", "cover"],
+        "general.artwork_cover" => &[
+            "hide cover",
+            "black background",
+            "backdrop",
+            "album art",
+            "visualizer only",
+        ],
         "general.slot_text_links" => &[
             "hyperlink",
             "clickable text",